:recycle: Move primitive utils
